Implement near lake indexer framework
Previous framework (near-indexer) was dropped in favor of near lake framework.
This new framework is less resource intensive to run and easier to set up.
It reads a data stream from an AWS S3 Lake instance and process it in the same way as the near-indexer does.

Change nft_staking contract back to previous version
Keeps tested use of near-sdk collections and implements view calls with serializable structs.

nearcore deleted